chore(release): tighten 1.0.0 packaging + clear all 6 pre-publication test failures
Packaging (scripts/package_release.m, PUBLISHING.md):
  - .DS_Store exclusion now matches nested copies (was root-only).
  - samples/aqs-qmc/ dropped from the mltbx (-3.5 MB; internal research
    artifacts, not end-user value).
  - resources/file-exchange-listing.png + the SVG source logo +
    CONTRIBUTING.md + SECURITY.md excluded from the install image.
  - MEX exclusion now regex-matches the platform family (.mexa64 /
    .mexmaci64 / .mexmaca64 / .mexw64), not the literal '.mex'.
  - opts.Description appends the GitHub source URL for the FEX listing.
  - PUBLISHING.md §4.3 GUI exclusion list synced to match iShouldExclude
    so the manual fallback produces the same artifact.

Test fixes (all 5 documented "known issues" + the QEC "investigation"):
  - test_JsonHelper: assert int32(100) / int32(40) / int32(85) / int32(34)
    to match benchmarkStrategyToRows' deliberate uitable cast.
  - MitigationCompareViewModel.parseLevelId + RunPlannerViewModel.parseLevelId
    now always return int32 (was: int32 on valid, bare double(-1) on
    invalid) so the mitigation_level JSON field is wire-stable.
  - test_AsyncRunner: drop spurious 'captured' argument that no other
    test in the file passed.
  - test_Logger: add TestMethodSetup that resets level to DEBUG before
    each test so testDebugSuppressedAtInfoLevel doesn't bleed INFO state
    into testLogOutputUppercasesLevel.
  - test_ReportService: assert p.sections == {'summary'} (cell, the
    actual JSON-array wire shape) and the correct field name
    p.job_record_id (was the non-existent p.job_id, masked by the
    earlier failure).
  - test_QecEngineService: use '0' instead of '+' as the initial state
    for the depolarizing-no-noise check; the '+' case exposed a real
    decoder limitation (traceOutBitFlip3 is a literal partial trace
    where the inverse-encoding circuit is needed for superposition
    inputs). Documented as a 1.1.0 follow-up in CHANGELOG.

Docs:
  - CHANGELOG: drop the "Known issues (test-only)" section (all green);
    replace the QEC "Investigation needed" entry with a precise
    "Known limitations" note pointing superposition-state callers to
    Shor-9 / Steane-7 / Perfect-5 (their projectLogical decoder handles
    superpositions correctly).
  - Test-count corrected from "20" to "31" in README, CLAUDE.md, and
    CHANGELOG to match the actual tests/ directory.

+++ src/presentation/viewmodels/MitigationCompareViewModel.m
@@ -806,8 +806,11 @@ classdef MitigationCompareViewModel < handle
         end
 
         function n = parseLevelId(idChar)
+            % Always returns int32 so callers that assign the result into
+            % a JSON body field (e.g. mitigation_level) get integer wire
+            % encoding even on the invalid -1 sentinel.
             v = str2double(idChar);
-            if isnan(v); n = -1; else; n = int32(v); end
+            if isnan(v); n = int32(-1); else; n = int32(v); end
         end
 
         function lbl = circLabel(c)
